+++ esmvalcore/cmor/_fixes/fix.py
@@ -57,7 +57,8 @@ class Fix:
             CMOR table entry of the variable.
         extra_facets:
             Extra facets are mainly used for data outside of the big projects
-            like CMIP, CORDEX, obs4MIPs. For details, see :ref:`extra_facets`.
+            like CMIP, CORDEX, obs4MIPs. For details, see
+            :ref:`config-extra-facets`.
         session:
             Current session which includes configuration and directory
             information.
@@ -223,7 +224,8 @@ class Fix:
             Variable's short name.
         extra_facets:
             Extra facets are mainly used for data outside of the big projects
-            like CMIP, CORDEX, obs4MIPs. For details, see :ref:`extra_facets`.
+            like CMIP, CORDEX, obs4MIPs. For details, see
+            :ref:`config-extra-facets`.
         session:
             Current session which includes configuration and directory
             information.
