+++ esmvalcore/config/_config_validators.py
@@ -7,6 +7,7 @@ import os.path
 import warnings
 from collections.abc import Callable, Iterable
 from functools import lru_cache, partial
+from importlib.resources import files as importlib_files
 from pathlib import Path
 from typing import Any
 
@@ -14,11 +15,7 @@ from packaging import version
 
 from esmvalcore import __version__ as current_version
 from esmvalcore.cmor.check import CheckLevels
-from esmvalcore.config._config import (
-    TASKSEP,
-    importlib_files,
-    load_config_developer,
-)
+from esmvalcore.config._config import TASKSEP, load_config_developer
 from esmvalcore.exceptions import (
     ESMValCoreDeprecationWarning,
     InvalidConfigParameter,
@@ -348,7 +345,7 @@ _validators = {
     "download_dir": validate_path,
     "drs": validate_drs,
     "exit_on_warning": validate_bool,
-    "extra_facets_dir": validate_extra_facets_dir,
+    "extra_facets": validate_dict,
     "log_level": validate_string,
     "logging": validate_dict,
     "max_datasets": validate_int_positive_or_none,
@@ -369,6 +366,8 @@ _validators = {
     # TODO: remove in v2.14.0
     # config location
     "config_file": validate_path,
+    # TODO: remove in v2.15.0
+    "extra_facets_dir": validate_extra_facets_dir,
 }
 
 
@@ -422,10 +421,35 @@ def deprecate_config_file(validated_config, value, validated_value):
     _handle_deprecation(option, deprecated_version, remove_version, more_info)
 
 
+# TODO: remove in v2.15.0
+def deprecate_extra_facets_dir(validated_config, value, validated_value):
+    """Deprecate ``extra_facets_dir`` option.
+
+    Parameters
+    ----------
+    validated_config: ValidatedConfig
+        ``ValidatedConfig`` instance which will be modified in place.
+    value: Any
+        Raw input value for ``extra_facets_dir`` option.
+    validated_value: Any
+        Validated value for ``extra_facets_dir`` option.
+
+    """
+    validated_config  # noqa: B018
+    value  # noqa: B018
+    validated_value  # noqa: B018
+    option = "extra_facets_dir"
+    deprecated_version = "2.13.0"
+    remove_version = "2.15.0"
+    more_info = " Please use the option `extra_facets` instead."
+    _handle_deprecation(option, deprecated_version, remove_version, more_info)
+
+
 # Example usage: see removed files in
 # https://github.com/ESMValGroup/ESMValCore/pull/2213
 _deprecators: dict[str, Callable] = {
     "config_file": deprecate_config_file,  # TODO: remove in v2.14.0
+    "extra_facets_dir": deprecate_extra_facets_dir,  # TODO: remove in v2.15.0
 }
 
 
@@ -434,4 +458,5 @@ _deprecators: dict[str, Callable] = {
 # https://github.com/ESMValGroup/ESMValCore/pull/2213
 _deprecated_options_defaults: dict[str, Any] = {
     "config_file": None,  # TODO: remove in v2.14.0
+    "extra_facets_dir": [],  # TODO: remove in v2.15.0
 }
