+++ esmvalcore/config/configurations/defaults/extra_facets_icon.yml
@@ -1,0 +1,146 @@
+# Extra facets for native ICON model output
+
+# Notes:
+# - All facets can also be specified in the recipes. The values given here are
+#   only defaults.
+# - The facet ``var_type`` has to be specified in the recipe if it is not given
+#   here and default DRS is used.
+
+# A complete list of supported keys is given in the documentation (see
+# ESMValCore/doc/quickstart/find_data.rst).
+---
+
+extra_facets:
+  ICON:
+    ICON:
+      '*':
+        # Cell measures
+        areacella: {raw_name: cell_area}
+        areacello: {raw_name: cell_area}
+
+        # 2D atmosphere variables
+        asr: {var_type: atm_2d_ml}
+        clivi: {var_type: atm_2d_ml}
+        clt: {var_type: atm_2d_ml}
+        clwvi: {var_type: atm_2d_ml}
+        evspsbl: {var_type: atm_2d_ml}
+        hfls: {var_type: atm_2d_ml}
+        hfns: {var_type: atm_2d_ml}
+        hfss: {var_type: atm_2d_ml}
+        lwcre: {var_type: atm_2d_ml}
+        lwp: {raw_name: cllvi, var_type: atm_2d_ml}
+        netcre: {var_type: atm_2d_ml}
+        pr: {var_type: atm_2d_ml}
+        prw: {var_type: atm_2d_ml}
+        ps: {var_type: atm_2d_ml}
+        psl: {var_type: atm_2d_ml}
+        rlds: {var_type: atm_2d_ml}
+        rldscs: {var_type: atm_2d_ml}
+        rlns: {var_type: atm_2d_ml}
+        rlus: {var_type: atm_2d_ml}
+        rlut: {var_type: atm_2d_ml}
+        rlutcs: {var_type: atm_2d_ml}
+        rsds: {var_type: atm_2d_ml}
+        rsdscs: {var_type: atm_2d_ml}
+        rsdt: {var_type: atm_2d_ml}
+        rsns: {var_type: atm_2d_ml}
+        rsnt: {var_type: atm_2d_ml}
+        rsus: {var_type: atm_2d_ml}
+        rsuscs: {var_type: atm_2d_ml}
+        rsut: {var_type: atm_2d_ml}
+        rsutcs: {var_type: atm_2d_ml}
+        rtmt: {var_type: atm_2d_ml}
+        rtnt: {var_type: atm_2d_ml}
+        swcre: {var_type: atm_2d_ml}
+        tas: {var_type: atm_2d_ml}
+        tasmax: {var_type: atm_2d_ml}
+        tasmin: {var_type: atm_2d_ml}
+        tauu: {var_type: atm_2d_ml}
+        tauv: {var_type: atm_2d_ml}
+        ts: {var_type: atm_2d_ml}
+        uas: {var_type: atm_2d_ml}
+        vas: {var_type: atm_2d_ml}
+
+        # 3D atmosphere variables
+        cl: {var_type: atm_3d_ml}
+        cli: {var_type: atm_3d_ml}
+        clw: {var_type: atm_3d_ml}
+        hur: {raw_units: '1', var_type: atm_3d_ml}
+        hus: {var_type: atm_3d_ml}
+        ta: {var_type: atm_3d_ml}
+        ua: {var_type: atm_3d_ml}
+        va: {var_type: atm_3d_ml}
+        wap: {var_type: atm_3d_ml}
+        zg: {var_type: atm_3d_ml}
+
+        # 2D ocean variables
+        siconc: {raw_name: sic, raw_units: '1', var_type: atm_2d_ml}
+        siconca: {raw_name: sic, raw_units: '1', var_type: atm_2d_ml}
+
+    ICON-XPP:
+      '*':
+        # 2D atmosphere variables
+        asr: {var_type: atm_2d_ml}
+        clivi: {raw_name: tqi_dia, var_type: atm_2d_ml}
+        clt: {raw_name: clct, var_type: atm_2d_ml}
+        clwvi: {var_type: atm_2d_ml}
+        evspsbl: {raw_name: qhfl_s, raw_units: kg m-2 s-1, var_type: atm_2d_ml}
+        hfls: {raw_name: lhfl_s, var_type: atm_2d_ml}
+        hfns: {var_type: atm_2d_ml}
+        hfss: {raw_name: shfl_s, var_type: atm_2d_ml}
+        lwcre: {var_type: atm_2d_ml}
+        lwp: {raw_name: tqc_dia, var_type: atm_2d_ml}
+        netcre: {var_type: atm_2d_ml}
+        pr: {raw_name: tot_prec_rate, var_type: atm_2d_ml}
+        prw: {raw_name: tqv_dia, var_type: atm_2d_ml}
+        ps: {raw_name: pres_sfc, var_type: atm_2d_ml}
+        psl: {raw_name: pres_msl, var_type: atm_2d_ml}
+        rlns: {raw_name: thb_s, var_type: atm_2d_ml}
+        rlut: {raw_name: thb_t, var_type: atm_2d_ml}
+        rlutcs: {raw_name: lwflx_up_clr, var_type: atm_2d_ml}
+        rsdt: {raw_name: sod_t, var_type: atm_2d_ml}
+        rsns: {raw_name: sob_s, var_type: atm_2d_ml}
+        rsnt: {raw_name: sob_t, var_type: atm_2d_ml}
+        rsus: {raw_name: sou_s, var_type: atm_2d_ml}
+        rsut: {raw_name: sou_t, var_type: atm_2d_ml}
+        rsutcs: {raw_name: swflx_up_clr, var_type: atm_2d_ml}
+        rtmt: {var_type: atm_2d_ml}
+        rtnt: {var_type: atm_2d_ml}
+        sfcWind: {raw_name: sp_10m, var_type: atm_2d_ml}
+        sftlf: {raw_name: fr_land, raw_units: '1', var_type: atm_2d_ml}
+        swcre: {var_type: atm_2d_ml}
+        tas: {raw_name: t_2m, var_type: atm_2d_ml}
+        tauu: {raw_name: umfl_s, var_type: atm_2d_ml}
+        tauv: {raw_name: vmfl_s, var_type: atm_2d_ml}
+        ts: {raw_name: t_s, var_type: atm_2d_ml}
+
+        # 3D atmosphere variables
+        cl: {raw_name: clc, var_type: atm_3d_ml}
+        cli: {raw_name: tot_qi_dia, var_type: atm_3d_ml}
+        clw: {raw_name: tot_qc_dia, var_type: atm_3d_ml}
+        hur: {raw_name: rh, var_type: atm_3d_ml}
+        hus: {raw_name: qv, var_type: atm_3d_ml}
+        ta: {raw_name: temp, var_type: atm_3d_ml}
+        ua: {raw_name: u, var_type: atm_3d_ml}
+        va: {raw_name: v, var_type: atm_3d_ml}
+        wap: {raw_name: omega, var_type: atm_3d_ml}
+        zg: {raw_name: geopot, var_type: atm_3d_ml}
+
+        # 2D land variables
+        gpp: {raw_name: assimi_gross_assimilation_box, var_type: jsb_2d_ml}
+        lai: {raw_name: pheno_lai_box, var_type: jsb_2d_ml, raw_units: '1'}
+
+        # 2D ocean variables
+        hfds: {raw_name: HeatFlux_Total, var_type: oce_dbg}
+        mlotst: {raw_name: mld, var_type: oce_dbg}
+        siconc: {raw_name: conc, var_type: oce_ice, raw_units: '1'}
+        siconca: {raw_name: fr_seaice, var_type: atm_2d_ml}
+        sithick: {raw_name: hi, var_type: oce_ice}
+        tos: {raw_name: t_seasfc, var_type: atm_2d_ml}
+        zos: {raw_name: zos, var_type: oce_dbg}
+
+        # 3D ocean variables
+        so: {raw_name: so, var_type: oce_def, raw_units: 0.001}
+        thetao: {raw_name: to, var_type: oce_def, raw_units: degC}
+        uo: {raw_name: u, var_type: oce_def}
+        vo: {raw_name: v, var_type: oce_def}
