--- esmvalcore/config/extra_facets/cmip5-fx.yml
@@ -1,5 +1,0 @@
----
-'*':
-  'fx':
-    '*':
-      ensemble: r0i0p0
