--- esmvalcore/config/extra_facets/cmip5-product.yml
@@ -1,5 +1,0 @@
----
-'*':
-  '*':
-    '*':
-      product: ['output1', 'output2']
