+++ tests/integration/cmor/_fixes/icon/test_icon.py
@@ -27,7 +27,7 @@ from esmvalcore.cmor._fixes.icon.icon import (
 from esmvalcore.cmor.fix import Fix
 from esmvalcore.cmor.table import CoordinateInfo, get_var_info
 from esmvalcore.config import CFG
-from esmvalcore.config._config import get_extra_facets
+from esmvalcore.config._extra_facets import get_extra_facets
 from esmvalcore.dataset import Dataset
 
 TEST_GRID_FILE_URI = (
@@ -173,7 +173,7 @@ def _get_fix(mip, short_name, fix_name, session=None):
         mip=mip,
         short_name=short_name,
     )
-    extra_facets = get_extra_facets(dataset, ())
+    extra_facets = get_extra_facets(dataset)
     extra_facets["frequency"] = "mon"
     extra_facets["exp"] = "amip"
     vardef = get_var_info(project="ICON", mip=mip, short_name=short_name)
